+++ src/Contracts/Requests/UpdateMasterProduct.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the UpdateMasterProduct interface.
+ *
+ * @copyright   Copyright (c) 2023 Attila Fulop
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-01-06
+ *
+ */
+
+namespace Vanilo\Admin\Contracts\Requests;
+
+use Konekt\Concord\Contracts\BaseRequest;
+
+interface UpdateMasterProduct extends BaseRequest
+{
+}